requirements.txt
appnope==0.1.0
botocore==1.5.65
decorator==4.0.11
docutils==0.13.1
grpcio==1.3.5
grpcio-tools==1.3.5
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jmespath==0.9.3
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.1
Pygments==2.2.0
pynamodb==3.0.1
python-dateutil==2.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
